various updates based on review from @to266
