+++ hyperspy/model.py
@@ -16,6 +16,7 @@
 # You should have received a copy of the GNU General Public License
 # along with  HyperSpy.  If not, see <http://www.gnu.org/licenses/>.
 
+import copy
 import os
 import tempfile
 import numbers
@@ -378,6 +379,99 @@ class BaseModel(list):
 
     __touch = _touch
 
+    def as_signal(self, component_list=None, out_of_range_to_nan=True,
+                  show_progressbar=None):
+        """Returns a recreation of the dataset using the model.
+        the spectral range that is not fitted is filled with nans.
+
+        Parameters
+        ----------
+        component_list : list of hyperspy components, optional
+            If a list of components is given, only the components given in the
+            list is used in making the returned spectrum. The components can
+            be specified by name, index or themselves.
+        out_of_range_to_nan : bool
+            If True the spectral range that is not fitted is filled with nans.
+        show_progressbar : None or bool
+            If True, display a progress bar. If None the default is set in
+            `preferences`.
+
+        Returns
+        -------
+        spectrum : An instance of the same class as `spectrum`.
+
+        Examples
+        --------
+        >>> s = hs.signals.Spectrum(np.random.random((10,100)))
+        >>> m = s.create_model()
+        >>> l1 = hs.model.components.Lorentzian()
+        >>> l2 = hs.model.components.Lorentzian()
+        >>> m.append(l1)
+        >>> m.append(l2)
+        >>> s1 = m.as_signal()
+        >>> s2 = m.as_signal(component_list=[l1])
+
+        """
+        # change actual values to whatever except bool
+        _multi_on_ = '_multi_on_'
+        _multi_off_ = '_multi_off_'
+        if show_progressbar is None:
+            show_progressbar = preferences.General.show_progressbar
+
+        if component_list:
+            component_list = [self._get_component(x) for x in component_list]
+            active_state = []
+            for component_ in self:
+                if component_.active_is_multidimensional:
+                    if component_ not in component_list:
+                        active_state.append(_multi_off_)
+                        component_._toggle_connect_active_array(False)
+                        component_.active = False
+                    else:
+                        active_state.append(_multi_on_)
+                else:
+                    active_state.append(component_.active)
+                    if component_ in component_list:
+                        component_.active = True
+                    else:
+                        component_.active = False
+        data = np.empty(self.signal.data.shape, dtype='float')
+        data.fill(np.nan)
+        if out_of_range_to_nan is True:
+            channel_switches_backup = copy.copy(self.channel_switches)
+            self.channel_switches[:] = True
+        maxval = self.axes_manager.navigation_size
+        pbar = progressbar.progressbar(maxval=maxval,
+                                       disabled=not show_progressbar)
+        i = 0
+        for index in self.axes_manager:
+            self.fetch_stored_values(only_fixed=False)
+            data[self.axes_manager._getitem_tuple][
+                self.channel_switches] = self.__call__(
+                non_convolved=not self.convolved, onlyactive=True)
+            i += 1
+            if maxval > 0:
+                pbar.update(i)
+        pbar.finish()
+        if out_of_range_to_nan is True:
+            self.channel_switches[:] = channel_switches_backup
+        signal = self.signal.__class__(
+            data,
+            axes=self.spectrum.axes_manager._get_axes_dicts())
+        signal.metadata.General.title = (
+            self.signal.metadata.General.title + " from fitted model")
+        signal.metadata.Signal.binned = self.signal.metadata.Signal.binned
+
+        if component_list:
+            for component_ in self:
+                active_s = active_state.pop(0)
+                if isinstance(active_s, bool):
+                    component_.active = active_s
+                else:
+                    if active_s == _multi_off_:
+                        component_._toggle_connect_active_array(True)
+        return signal
+
     @property
     def _plot_active(self):
         if self._plot is not None and self._plot.is_active() is True:
@@ -551,7 +645,7 @@ class BaseModel(list):
     def red_chisq(self):
         """Reduced chi-squared. Calculated from self.chisq and self.dof
         """
-        tmp = self.chisq / (- self.dof + sum(self.channel_switches) - 1)
+        tmp = self.chisq / (- self.dof + self.channel_switches.sum() - 1)
         tmp.metadata.General.title = self.signal.metadata.General.title + \
             ' reduced chi-squared'
         return tmp
@@ -974,105 +1068,6 @@ class BaseModel(list):
 
         self.fetch_stored_values()
 
-    def plot(self, plot_components=False):
-        """Plots the current spectrum to the screen and a map with a
-        cursor to explore the SI.
-
-        Parameters
-        ----------
-        plot_components : bool
-            If True, add a line per component to the signal figure.
-
-        """
-
-        # If new coordinates are assigned
-        self.spectrum.plot()
-        _plot = self.spectrum._plot
-        l1 = _plot.signal_plot.ax_lines[0]
-        color = l1.line.get_color()
-        l1.set_line_properties(color=color, type='scatter')
-
-        l2 = hyperspy.drawing.spectrum.SpectrumLine()
-        l2.data_function = self._model2plot
-        l2.set_line_properties(color='blue', type='line')
-        # Add the line to the figure
-        _plot.signal_plot.add_line(l2)
-        l2.plot()
-        on_figure_window_close(_plot.signal_plot.figure,
-                               self._close_plot)
-
-        self._model_line = l2
-        self._plot = self.spectrum._plot
-        self._connect_parameters2update_plot()
-        if plot_components is True:
-            self.enable_plot_components()
-
-    @staticmethod
-    def _connect_component_line(component):
-        if hasattr(component, "_model_plot_line"):
-            component.connect(component._model_plot_line.update)
-            for parameter in component.parameters:
-                parameter.connect(component._model_plot_line.update)
-
-    @staticmethod
-    def _disconnect_component_line(component):
-        if hasattr(component, "_model_plot_line"):
-            component.disconnect(component._model_plot_line.update)
-            for parameter in component.parameters:
-                parameter.disconnect(component._model_plot_line.update)
-
-    def _connect_component_lines(self):
-        for component in [component for component in self if
-                          component.active]:
-            self._connect_component_line(component)
-
-    def _disconnect_component_lines(self):
-        for component in [component for component in self if
-                          component.active]:
-            self._disconnect_component_line(component)
-
-    def _plot_component(self, component):
-        line = hyperspy.drawing.spectrum.SpectrumLine()
-        line.data_function = component._component2plot
-        # Add the line to the figure
-        self._plot.signal_plot.add_line(line)
-        line.plot()
-        component._model_plot_line = line
-        self._connect_component_line(component)
-
-    @staticmethod
-    def _update_component_line(component):
-        if hasattr(component, "_model_plot_line"):
-            component._model_plot_line.update()
-
-    def _disable_plot_component(self, component):
-        self._disconnect_component_line(component)
-        if hasattr(component, "_model_plot_line"):
-            component._model_plot_line.close()
-            del component._model_plot_line
-        self._plot_components = False
-
-    def _close_plot(self):
-        if self._plot_components is True:
-            self.disable_plot_components()
-        self._disconnect_parameters2update_plot()
-        self._model_line = None
-
-    def enable_plot_components(self):
-        if self._plot is None or self._plot_components:
-            return
-        self._plot_components = True
-        for component in [component for component in self if
-                          component.active]:
-            self._plot_component(component)
-
-    def disable_plot_components(self):
-        if self._plot is None:
-            return
-        for component in self:
-            self._disable_plot_component(component)
-        self._plot_components = False
-
     def assign_current_values_to_all(self, components_list=None, mask=None):
         """Set parameter values for all positions to the current ones.
 
@@ -1196,173 +1191,6 @@ class BaseModel(list):
                     if not hasattr(parameter.value, '__iter__'):
                         print("\t\t%s\t%g" % (
                             parameter.name, parameter.value))
-
-    def enable_adjust_position(
-            self, components=None, fix_them=True, show_label=True):
-        """Allow changing the *x* position of component by dragging
-        a vertical line that is plotted in the signal model figure
-
-        Parameters
-        ----------
-        components : {None, list of components}
-            If None, the position of all the active components of the
-            model that has a well defined *x* position with a value
-            in the axis range will get a position adjustment line.
-            Otherwise the feature is added only to the given components.
-            The components can be specified by name, index or themselves.
-        fix_them : bool
-            If True the position parameter of the components will be
-            temporarily fixed until adjust position is disable.
-            This can
-            be useful to iteratively adjust the component positions and
-            fit the model.
-        show_label : bool, optional
-            If True, a label showing the component name is added to the
-            plot next to the vertical line.
-
-        See also
-        --------
-        disable_adjust_position
-
-        """
-        if (self._plot is None or
-                self._plot.is_active() is False):
-            self.plot()
-        if self._position_widgets:
-            self.disable_adjust_position()
-        on_figure_window_close(self._plot.signal_plot.figure,
-                               self.disable_adjust_position)
-        if components:
-            components = [self._get_component(x) for x in components]
-        else:
-            self._adjust_position_all = (fix_them, show_label)
-
-        components = components if components else self
-        if not components:
-            # The model does not have components so we do nothing
-            return
-        components = [
-            component for component in components if component.active]
-        for component in components:
-            self._make_position_adjuster(component, fix_them, show_label)
-
-    def _make_position_adjuster(self, component, fix_it, show_label):
-        if (component._position is not None and
-                not component._position.twin):
-            set_value = component._position._set_value
-            get_value = component._position._get_value
-        else:
-            return
-        # Create an AxesManager for the widget
-        axis_dict = self.axes_manager.signal_axes[0].get_axis_dictionary()
-        am = AxesManager([axis_dict, ])
-        am._axes[0].navigate = True
-        try:
-            am._axes[0].value = get_value()
-        except TraitError:
-            # The value is outside of the axis range
-            return
-        # Create the vertical line and labels
-        if show_label:
-            self._position_widgets.extend((
-                DraggableVerticalLine(am),
-                DraggableLabel(am),))
-            # Store the component for bookkeeping, and to reset
-            # its twin when disabling adjust position
-            self._position_widgets[-2].component = component
-            self._position_widgets[-1].component = component
-            w = self._position_widgets[-1]
-            w.string = component._get_short_description().replace(
-                ' component', '')
-            w.set_mpl_ax(self._plot.signal_plot.ax)
-            self._position_widgets[-2].set_mpl_ax(
-                self._plot.signal_plot.ax)
-        else:
-            self._position_widgets.extend((
-                DraggableVerticalLine(am),))
-            # Store the component for bookkeeping, and to reset
-            # its twin when disabling adjust position
-            self._position_widgets[-1].component = component
-            self._position_widgets[-1].set_mpl_ax(
-                self._plot.signal_plot.ax)
-        # Create widget -> parameter connection
-        am._axes[0].continuous_value = True
-        am._axes[0].on_trait_change(set_value, 'value')
-        # Create parameter -> widget connection
-        # This is done with a duck typing trick
-        # We disguise the AxesManager axis of Parameter by adding
-        # the _twin attribute
-        am._axes[0]._twins = set()
-        component._position.twin = am._axes[0]
-
-    def disable_adjust_position(self):
-        """Disables the interactive adjust position feature
-
-        See also
-        --------
-        enable_adjust_position
-
-        """
-        self._adjust_position_all = False
-        while self._position_widgets:
-            pw = self._position_widgets.pop()
-            if hasattr(pw, 'component'):
-                pw.component._position.twin = None
-                del pw.component
-            pw.close()
-            del pw
-
-    def fit_component(
-            self,
-            component,
-            signal_range="interactive",
-            estimate_parameters=True,
-            fit_independent=False,
-            only_current=True,
-            **kwargs):
-        """Fit just the given component in the given signal range.
-
-        This method is useful to obtain starting parameters for the
-        components. Any keyword arguments are passed to the fit method.
-
-        Parameters
-        ----------
-        component : component instance
-            The component must be in the model, otherwise an exception
-            is raised. The component can be specified by name, index or itself.
-        signal_range : {'interactive', (left_value, right_value), None}
-            If 'interactive' the signal range is selected using the span
-             selector on the spectrum plot. The signal range can also
-             be manually specified by passing a tuple of floats. If None
-             the current signal range is used.
-        estimate_parameters : bool, default True
-            If True will check if the component has an
-            estimate_parameters function, and use it to estimate the
-            parameters in the component.
-        fit_independent : bool, default False
-            If True, all other components are disabled. If False, all other
-            component paramemeters are fixed.
-
-        Examples
-        --------
-        Signal range set interactivly
-
-        >>> g1 = hs.model.components.Gaussian()
-        >>> m.append(g1)
-        >>> m.fit_component(g1)
-
-        Signal range set through direct input
-
-        >>> m.fit_component(g1, signal_range=(50,100))
-        """
-        component = self._get_component(component)
-        cf = ComponentFit(self, component, signal_range,
-                          estimate_parameters, fit_independent,
-                          only_current, **kwargs)
-        if signal_range == "interactive":
-            cf.edit_traits()
-        else:
-            cf.apply()
 
     def set_parameters_not_free(self, component_list=None,
                                 parameter_name_list=None):
